Extend GPR820 support to OCaml 4.00.x
4.00.0 is when the -with-debug-runtime switch was added to configure, so
this extends the fixes from GPR820 back to that version.

Although there's evidence in 3.07-3.09.3 that building ocamlrund.exe was
on someone's mind, proper support wasn't added to the build system until
the 64-bit Windows port was added in 3.10.0 (when build instructions
appear in byterun/Makefile.nt). However, as the ability to switch
runtime wasn't added until 4.00.0, it doesn't seem worth back-porting
any of the fixes to 3.10, 3.11 and 3.12.